leaf51: add loc guard and modularize engine crate
